Extract new WooPlan to its own file
